Fix advert with no name
It is possible, though not very useful, to send an advert with no name field.
If one was recieved, it would cause an exception when passing the advert
details to the app.

The name of adverts with no name field is now set to "Unnamed xxxxxxxx",
plus a red question mark emoji for the display